# generator-jhipster-go/generators/server/templates/go/go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample service for managing events",
        "title": "Events API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "soberkoder@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7090",
    "basePath": "/",
    "paths": {
        "/delete/{id}": {
            "delete": {
                "description": "Delete the event by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Delete an event By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delete event",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/event": {
            "post": {
                "description": "Create a new event with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create a new event",
                "parameters": [
                    {
                        "description": "Create event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.Event"
                        }
                    }
                }
            }
        },
        "/event/{id}": {
            "get": {
                "description": "Fetch the event details by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Fetch an event By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Get event",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.Event"
                        }
                    }
                }
            }
        },
        "/events": {
            "get": {
                "description": "Get details of all events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get details of all events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.Event"
                            }
                        }
                    }
                }
            }
        },
        "/update": {
            "patch": {
                "description": "Updates an event with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Update an event",
                "parameters": [
                    {
                        "description": "Update event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.Event"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Update event",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "services.Event": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}